backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
channels>=4.0
drf-spectacular>=0.27
django-filter>=24.2
django-environ>=0.11
celery>=5.4
redis>=5.0
psycopg2-binary>=2.9
django-auditlog>=2.3
nltk>=3.9
djangorestframework-simplejwt>=5.3
pyotp>=2.9
SpeechRecognition>=3.10
pydub>=0.25
